Fix error in setting change callback and clean up code
Main is undefined and Main.refresh was being called whenever a setting was
changed, causing an error. This was unneeded anyway because there is a signal
connected in anyway in extension.js to call refresh. Remove call to
Main.refresh.
Clean up code, sticking as closely as possible to the 80 character rule.