Add icon hiding functionality
Icons can be hidden by their WM class using the extension settings. The UI (in
the settings) needs a polish.

The hidden icons are saved into GSettings (hidden-icons) whenever an icon's
toggle is clicked, and there is a signal connected in the main extension. This
then calls the refresh function, which fetches the updated GSettings values, and
then iterates over the icons and either shows or hides them depending on whether
the icon's WM class appears in the hidden icons array.