Hidden icons feature (#6)
* Change defaults and add hidden-icons key

hidden-icons is an array of strings, and will store the WM classes of icons to
be hidden.

* Add signal for hidden-icons and clean up WM class check

Whenever hidden-icons is changed, the refresh function is called. This function
fetches the updated settings and applies them. The hidden-icons setting is
applied by calling onTrayIconRemoved on the icon.
getWmClass simply gets an icon's WM class.

* Add current-icons array to hold current icons

* Add icon hiding functionality

Icons can be hidden by their WM class using the extension settings. The UI (in
the settings) needs a polish.

The hidden icons are saved into GSettings (hidden-icons) whenever an icon's
toggle is clicked, and there is a signal connected in the main extension. This
then calls the refresh function, which fetches the updated GSettings values, and
then iterates over the icons and either shows or hides them depending on whether
the icon's WM class appears in the hidden icons array.